version 0.07
- User,Post, Comment ID 타입 Long으로 수정 (관련 메서드 및 이벤트 수정)
- User avatar -> imageUrl 변경

- Ktor API 설정
 1. JWT 토큰 헤더 추가
 2. Multipart 설정 추가

- 팔로우 처리
 1. 팔로우/언팔로우 기능 구현
 2. 팔로우 추천 유저 목록 가져오기

- 게시물
 1. 게시물 목록 가져오기 (+특정 유저 게시물 목록)
 2. 좋아요 / 좋아요 해제
 3. 게시물 상세 정보 요청
 4. 게시물 작성

+++ androidApp/src/main/java/com/skymilk/socialapp/android/presentation/screen/main/follows/FollowsScreen.kt
@@ -16,7 +16,7 @@ fun FollowsScreen(
     modifier: Modifier = Modifier,
     followsUsersState: FollowsState,
     onEvent: (FollowsEvent) -> Unit,
-    onNavigateToProfile: (Int) -> Unit
+    onNavigateToProfile: (Long) -> Unit
 ) {
     Box(
         modifier = modifier.fillMaxSize(),
