+++ androidApp/src/main/java/com/skymilk/socialapp/android/presentation/screen/main/profile/ProfileViewModel.kt
@@ -17,7 +17,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 class ProfileViewModel(
-    private val userId: Int
+    private val userId: Long
 ): ViewModel() {
 
     private val _profileState = MutableStateFlow<ProfileState>(ProfileState.Initial)
